modified

+++ src/main/java/com/isoft/community/dto/CommentDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CommentDTO {
-    private long parent_id;
+    private Integer parent_id;
     private String content;
     private Integer type;
 }
